Replace coc.nvim with ALE/vim-lsp/etc
So far, using `coc.nvim` and associated plugins wasn't working very well
and also wasn't particuarly very ergonomic to work with.

This replaces that with ALE (what was used before), adds `vim-lsp` (with
`vim-lsp-settings` to make configuration easier) and introduces
`asyncomplete` for autocompletion, a setup which seems to work quite
well.

So far, the only manual setup has been needed for Ruby, where
`solargraph` is preferred.

https://github.com/dense-analysis/ale
https://github.com/prabirshrestha/vim-lsp
https://github.com/mattn/vim-lsp-settings
https://github.com/prabirshrestha/asyncomplete.vim