Add a command to diff against production
By convention, most projects follow the pattern of having a branch
called `production` which when pushed to triggers a workflow that causes
a deployment.

It's very helpful to know what you might be deploying — or to check if a
deployment should be done (if it's manual) so this command provides a
nice short version to show which commits are missing from `production` but
are on `main`.

This is mostly taken from the `gll` and `gl` aliases, but with a much
shorter output that could be shared without extraneous information.