Add vale for linting words
Vale brings editorial style guides as a linting tool. This adds Vale to
vim, via ALE and introduces the `proselint`, `write-good` and `alex`
style guides to get started with. These cover common writing mistakes
and also uninclusive language which is nice to catch early.

To use this, we need to run `vale sync` to bring in the packages, which
are installed in `~/.config/vale`.

https://vale.sh
https://github.com/dense-analysis/ale/issues/642
https://til.codeinthehole.com/posts/how-to-get-vale-to-work-with-vims-ale-plugin/