Add a script to help bump bundler
This script bumps the version of bundler used in a project (based on the
argument given) by first bundling in a container and then updating the
version. If there's a change, it then creates a branch, commits the
change and opens a PR on the current project for it.

The intention is to speed up bumping the Bundler version across a lot of
projects and it works well for this.