Add a script to help bump Ruby
This script bumps the version of Ruby currently used on a project (in
the `Gemfile`), along with the `.ruby-version`, and `.tool-versions` (if
either are present).

It does this by parsing the existing files, then updating the versions
in place (for `.ruby-version` and `.tool-versions`) or in a container
(for `Gemfile`) and then committing and opening a PR for it.

The intention is to speed up upgrading Ruby across lots of projects in
one go.