Add find-in-path
This finds items matching a pattern in the current `$PATH`. It could be
used to fuzzy match what's available, debug what is being called, or be
used to build out functionality like Git has - allowing you to extend
the command's `git` can run by having something in the `$PATH` which
matches `git-`.