Rebuild the Administrate demo creator
This has been broken since we changed Suspenders to be based around
generators, which invalidated the approach this was taking.

Here, we rename the command to start with `create-` which fits better,
adjust to use Suspenders via a generator (using `main`, as we need
unreleased support for Rails 8.0), fix an issue with GPG signing and
ignore a shellcheck error which actually works fine.

https://github.com/thoughtbot/suspenders/releases/tag/v20240516.0