Add ruby/node versions to the prompt
When jumping around lots of different projects, sometimes Ruby or Node
versions aren't detected correctly (or at all), and it's necessary to
check. This also helps with an at a glance view of whether or not we
should upgrade the language version if it's quite old.

It was inspired by a pattern used in the spacezsh-prompt project, which
helped a lot!

https://github.com/laggardkernel/spacezsh-prompt/blob/master/sections/sz%3A%3Aruby